auh tasks

+++ src/Routes/task.js
@@ -3,12 +3,13 @@ const mongoose =require('mongoose');
 
 const titleDesc = express.Router();
 const taskSchema = require('../db/models/task');
-
+const auth =require('../middleware/auth');
 //Save Task To DataBase
-titleDesc.post('/', (req, res, next) => {
+titleDesc.post('/', auth,(req, res, next) => {
     const task = new taskSchema({
         Title: req.body.title,
-        Description: req.body.desc
+        Description: req.body.desc,
+        owner :req.user._id
     })
     task.save().then(result => {
         res.status(200).json({
@@ -23,8 +24,8 @@ titleDesc.post('/', (req, res, next) => {
         })
 })
 // fetch Tasks From DataBase
-titleDesc.get('/', (req, res, next) => {
-    taskSchema.find().exec().then(data => {
+titleDesc.get('/',auth, (req, res, next) => {
+    taskSchema.find({owner:req.user._id}).exec().then(data => {
         res.status(200).json({
             message: "Sucess",
             dat: data
@@ -38,9 +39,9 @@ titleDesc.get('/', (req, res, next) => {
 })
 
 //fecth Single Task From DataBase
-titleDesc.get('/:id', (req, res, next) => {
-    const id = req.params.id
-    taskSchema.findById(id).exec().then(task => {
+titleDesc.get('/:id',auth, (req, res, next) => {
+    const _id = req.params.id
+    taskSchema.findOne({_id,owner:req.user._id}).exec().then(task => {
         if (!task) {
              res.status(404).json({
                 message: "Not Found"
@@ -66,7 +67,7 @@ titleDesc.patch('/:id', (req, res, next) => {
         Title: req.body.title,
         Description: req.body.desc
     }
-    taskSchema.updateOne({_id:id},req.body,{ runValidators: true }).exec().then(result => {
+    taskSchema.updateOne({_id:id,owner:req.user._id},req.body,{ runValidators: true }).exec().then(result => {
         res.status(200).json({
             message: "sucess",
             dat: result
@@ -80,11 +81,11 @@ titleDesc.patch('/:id', (req, res, next) => {
 })
 
 //Delete Single Task
-titleDesc.delete('/:id',(req,res,next)=>{
+titleDesc.delete('/:id',auth,(req,res,next)=>{
   const idd =req.params.id;
   var id = mongoose.Types.ObjectId(idd);
 
-  taskSchema.deleteOne({_id:id}).exec().then(result=>{
+  taskSchema.deleteOne({_id:id,owner:req.user._id}).exec().then(result=>{
     res.status(200).json({
         message:"Sucess",
         resu:result
